update: rename types files
