+++ src/data/token/token.repository.ts
@@ -5,7 +5,7 @@ import Context, { RequestContext } from 'bases/context';
 import env from 'env';
 
 import TokenHolder from './token.holder';
-import { Token, TokenContent } from './token.types';
+import { Token, TokenContent } from './token';
 
 // Repository
 class TokenRepository<T extends TokenHolder = TokenHolder> {
